+++ scripts/lichessTools-init.js
@@ -58,9 +58,11 @@ lichessTools.loadTool(LiChessTools.Tools.StickyStudySettings);
 lichessTools.loadTool(LiChessTools.Tools.ChapterInsert);
 lichessTools.loadTool(LiChessTools.Tools.ChapterPgnArea);
 lichessTools.loadTool(LiChessTools.Tools.MoveListOptions);
+lichessTools.loadTool(LiChessTools.Tools.StreamerTv);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);
+lichessTools.loadTool(LiChessTools.Tools.CopyPgnCommand);
 
 lichessTools.loadTool(LiChessTools.Tools.UserTvHistory);